Fix for Solaris/x86. With great help of Alexey Stroganov.
Open issue noted by Alexey:
- on Solaris dirent structure differs from Linux one so we may want 
  change test for directory on Solaris with following code:

  struct stat *s;
  stat(dir, s);
  if (s->st_mode & S_IFDIR) 
    return true;

+++ storage/falcon/Thread.cpp
@@ -245,7 +245,7 @@ Thread* Thread::getThread(const char *desc)
 
 #ifdef _PTHREADS
 	if (!thread->threadId)
-		thread->threadId = (unsigned long) pthread_self();
+		thread->threadId = pthread_self();
 #endif
 
 	return thread;
